requirements.txt
absl-py==0.8.0
amqp==2.5.1
asn1crypto==1.2.0
astor==0.8.0
billiard==3.6.1.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cryptography==2.8
cycler==0.10.0
cytoolz==0.10.0
dask==2.6.0
decorator==4.4.1
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.16.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.5
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
mtcnn==0.0.9
networkx==2.4
numpy==1.17.3
oauthlib==3.1.0
olefile==0.46
pandas==0.25.2
Pillow==6.2.0
protobuf==3.9.2
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.3
urllib3==1.24.2
vine==1.3.0
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
